pyproject.toml
[project]
name = "paper-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "playwright>=1.0",
    "Pillow>=10.5.0",
]
